refactored requests api module
